requirements.txt
apache-airflow-providers-snowflake
pandas
openpyxl
python-dotenv
snowflake-connector-python
snowflake-sqlalchemy